refactor: move landing sorting.rs inline tests to sibling tests.rs
Completes #862: the last of the 13 files listed in the issue.
frontend/src/pages/landing/sorting.rs was excluded from the earlier
sweep (#1066) to avoid a merge conflict with #1059, which has since
merged. Pure relocation via `use super::*;`, no logic or assertion
changes.
